✨ Feat: 회원 탈퇴 기능 구현

See merge request s12-final/S12P31C208!10

+++ backend/sleephony/src/main/java/com/c208/sleephony/domain/user/dto/response/GetUserProfileResponse.java
@@ -1,7 +1,6 @@
 package com.c208.sleephony.domain.user.dto.response;
 
 import lombok.*;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,8 +12,8 @@ import java.time.LocalDate;
 public class GetUserProfileResponse {
     private String email;
     private String nickname;
-    private BigDecimal height;
-    private BigDecimal weight;
+    private float height;
+    private float weight;
     private LocalDate birthDate;
     private String gender;
 }
